requirements.txt
PyQt6
psycopg[binary]
bcrypt
python-dotenv
dataclasses-json